requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
langchain-text-splitters>=0.0.1
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0
streamlit>=1.29.0
python-dotenv>=1.0.0
pypdf>=4.0.0
httpx>=0.26.0
